Moving populate published status to testing utils
